+++ views/reports.anomaly.csv.php
@@ -1,0 +1,8 @@
+<?php
+
+/**
+ * @var CView $this
+ * @var array $data
+ */
+
+echo zbx_toCSV(array_merge([$data['export_headers']], $data['export_rows']));